feat(structure): use individual finger controller components
BREAKING CHANGE: This commit reworks how the hand controlling is done.
Instead of a global hand controller that has multiple embedded finger
controllers, now it has a single Finger Controller MonoBehaviour that
is independent per finger and then a HandFacade that provides a simple
way of accessing the FingerController collection.

This means there doesn't need to be a bunch of repetitive methods for
setting and clearing the FingerController properties as the
FingerController can now be directly accessed. It also breaks down
the logic into the smallest chunk its actually doing, which is simply
controlling the finger.