feat(FingerController): provide threshold to determine lerp transition
Previously the lerped transition would only happen for the boolean or
override value and the float value would never lerp. This has now been
customised further so it is possible to lerp a float value if the
current curl value and the given target curl value exceed the
transition threshold.